New download page.
As opposed to being dynamically generated server-side (which got a bit
messy), this page will be dynamically updated using JavaScript with
humans taking back control over the output of the page.

It's now a bit easier to update links and the presentation separately.

Thanks to @RobertoPerico for the mascot graphics.